Clear Game, Vanilla Mode, Friends Mode - Added a clear method when exiting gameScreen. - Vanilla and Friends Pong selection: The variable to hold the option is game_mode.  The options are 'vanilla' and 'friends'. - AI Mode: The variable is ai_lvl and the options are 'easy' and 'hard'. - Friends Mode: It will update the game_mode variable to 'solo' or 'tourney'
The variables will be passed to Game constructor to decide the websocket URL. It will be used in the BPSlaveMaster to determine which game mode is choosen.

The ai_lvl options will be passed to the BPSlave directly via websocket to determine which AI difficulty to use. This variable will not be passed through BPSlaveMaster.

Not yet done: Timer countdown, player names, improvement on game front end, and other features.